+++ crates/common/src/grid/grids/grid_3d.rs
@@ -191,9 +191,7 @@ impl<T: GridParam, const LAYER_COUNT: usize> Grid3d<T, LAYER_COUNT> {
     }
 
     #[inline]
-    pub fn get_idx_unchecked(&self, point: impl GridPoint) -> usize {
-        point.as_index_unchecked(self.width() as usize)
-    }
+    pub fn get_idx_unchecked(&self, point: impl GridPoint) -> usize { point.as_index_unchecked(self.width()) }
 
     #[inline]
     pub fn index_to_pt(&self, idx: usize) -> Option<UVec2> {
